Introduce initial version of the crate
This change introduces the initial version of a crate that allows for
running Rust tests (functions annotated with the #[test] attribute) with
env_logger functionality initialized. Without that, each and every test
has to include setup code to initialize logging of on its own.

The crate is implemented as a procedural macro, which allows us to
define a custom attribute that users can then use similar to what they
already know -- all the heavy lifting happens transparently.