Allow overlaying of built-in #[test] attribute
Our crate has one problem that decreases its utility by a quite bit: the
user cannot import the crate provide test attribute under the name
'test' itself. Doing so will result in an endless recursion because we
emit code that uses #[test] that we expect to map to the built-in test
attribute, but that would then be overlaid with our custom one.

This change fixes the problem with some elaborate nesting of modules in
conjunction with nifty type aliasing. The reason for such an intricate
construct is that there seemingly is no way to reference a built-in
attribute once it has been overlaid by a user defined one.