Add more instructions relating to test output and log level  to README
This change adds instructions on how to retrieve the output of the tests
when using 'cargo test' to the README, as well as some information about
how to change the log level. While generally known, there is no good
reason to stop the explanation just before the output is actually
visible.