Add support for async attributes
This change introduces support for async functions to the wrapper
expansion logic. Note that in and of itself this logic is not useful
because test functions cannot really be async. However, the tokio crate
(and potentially others) add such support through a procedural macro. In
a future change we want to support "inner" test attributes (such as
tokio's), in which case we need to be able to properly handle async
functions.