Add support for inner test attributes
Conceptually, the test attribute provided by this crate is just a tiny
wrapper around the actual function that performs some initialization
before invoking it. We can very well envision having multiple
initialization steps, which could be accomplished by nesting #[test]
attributes.
This patch adds support for such nesting. Basically, we enhance the
procedural macro with an optional argument that, if specified, is used
as the inner #[test] attribute.
In conjunction with the previously added support for async functions, it
is now possible to use this test attribute in conjunction with that of
tokio. E.g.,

  #[test_env_log::test(tokio::test)]
  async fn with_tokio() {
    assert_eq!(async { 42 }.await, 42)
  }