Drop dependency to env_logger
From the looks of it, proc-macro crates' dependencies are treated a bit
differently by Cargo than those of library crates. In particular, they
are not "inherited" by (and, hence, available in) the consuming crate.
As such, the dependency to env_logger that we had specified all along
was effectively unused: while we emitted code that requires the
dependency to be available, it still had to be specified explicitly in
the crate making use of test-env-log.
This change fixes the confusion by converting the dependency to a
dev-only one. We require still require it for testing because our tests
execute the emitted code.