Add support for initializing tracing
There is a second kind of infrastructure in addition to logging that
requires initialization: tracing. In a nutshell, in order to collect
traces we need to register a subscriber. In our case, we want this
subscriber to log events and spans as it would do for regular log based
messages.
This change adds the infrastructure for doing so. In binaries we would
typically just set a global subscriber and be done. However, we want to
preserve per-test configurability and so we just set the subscriber for
the context of the test.
Because conceptually logging and tracing are two separate concerns and
applications are free to mix and match, we introduce two features, log
and trace, that enable one or the other or both.