Use default features for tracing dev-dependency
With this change we switch over to using the tracing dev-dependency with
the default features. In the past we have used no features, but it
turned out that some of our dependencies implicitly enabled the std and
attributes features.
By switching to the default set of features we hopefully are somewhat
more future proof.