Initialize tracing globally
So far we initialized tracing using the subscriber::set_default
functionality. With it, we set the subscriber for the test in question
and nowhere else. The thinking was that this is sufficient: tracing
should take care of using the subscriber in whatever threads the
corresponding task is migrated to. While it has not been verified that
this is indeed the behavior exhibited, it certainly is possible for
users to spin up a dedicated thread from a test and expect traces to be
captured for it as well. That, however, is not the case, because the
subscriber is thread-local.
With this change we switch over to using subscriber::set_global_default,
which performs tracing initialization globally. In so doing we
circumvent such potential problems. We also potentially gain a set of
new ones, for clients that expect tracing to be initialized only in
tests tagged with our #[test] attribute, but given that this is an
inherent design choice already made by the log crate, it seems sensible
to make tracing in this crate behave in the same fashion.