Eliminate need for emitting of ::f test function
When we devised the procedural macro we needed a way to reference the
default #[test] attribute. Because no way was found to do so, we had to
jump through a lot of hoops to get the same result: introduce a custom
module, emit a test function, f, mirroring the original test function
signature through an alias and super import, and then invoke a function
containing the original test code. As it turns out the default test
attribute is available as ::core::prelude::v1::test.
With this change we adjust the crate to use this name, removing all the
complex alias and module logic and no longer messing with test names in
the process.