Capture trace emitted log messages
By default, tracing messages go to standard output stream and as such
are not captured by libtest (used when running tests; which only
captures output from the print! family of macros). That means that
output is emitted directly, without the need to run the test with the
--nocapture option.
That is in contrast to the behavior of env_logger, which requires this
switch, because the log output is captured by default.
As it turns out since version 0.2.12 the tracing-subscriber crate
supports the capturing of trace output when running in a libtest
environment as well.
With this change we enable this usage, making behavior with the two
features we support more similar.