Capture `trace` logs with `libtest`
Configures the tracing subscriber to allow log messages to be captured
by `libtest`. This makes it a bit clearer whether log messages are in
the scope of your test when multiple tests are running for async tests.

As the method on `tracing_subscriber::fmt::SubscriberBuilder` was added
in `0.2.12`, this commit also updates the README and the changelog to
reflect that change.