Require tracing-subscriber in version 0.2.17
As it turns out, the recent addition of support for span events was only
introduced to the tracing-subscriber crate in version 0.2.17, meaning
that usage of earlier versions would cause a compilation error.
This change updates the README to explicitly state this requirement, and
bumps our dev-dependency to be in-line with it.