Recommend tracing-subscriber version 0.3
The tracing-subscriber crate has released a new breaking version. With
this change we update our dev-dependency accordingly and recommend users
use that. However, other than a making recommendation we have no
influence on which version client code actually consumes and given the
limited feature set that we use from the crate, both 0.2.17 and 0.3*
work just fine.

Signed-off-by: Sean Pianka <pianka@eml.cc>