Rename crate to test-log
This change renames the crate to test-log. The former name test-env-log
is no longer particularly appropriate, as it suggests an exclusive
relationship with env_logger, which in times past was the only supported
kind of test initialization. By now, however, we support tracing as
well.
The new name is shorter, more general, and arguably not suggesting
inappropriate connections to other crates.