Add CI job for running cargo doc
While we can't verify it completely, we at least want to make sure that
our documentation is not broken whenever changes are made. To that end,
this change introduces a new CI job that runs 'cargo doc' on the crate.
To make this check more effective, we also enable two more crate-wide
lints enforcing higher standards.