Enable extra-traits feature on syn dependency
As it turns out a Debug implementation for the syn::NestedMeta type is
only available if the 'extra-traits' feature of the crate is enabled.
That is implicitly the case for edition 2018 projects, where resolver v1
is used, which unifies features between regular and dev-dependencies.
However, if the crate is used from an edition 2021 project, resolver v2
is active and it will compile syn with only the explicitly specified
features -- which do not include 'extra-traits'. This may cause a build
failure.
This change fixes the problem by explicitly enabling the 'extra-traits'
feature. Unfortunately, we do not yet want to switch to using edition
2021 or resolver v2 for this crate ourselves -- which would allow us to
detect problems like this in CI moving forward -- as that involves yet
another (arguably significant) minimum supported Rust version bump. So
for now this fix is all we have.