Isolate generated initialization code from client code
The initialization code that we generate as part of the procedural macro
effectively lives in the same namespace (for values, but also types) as
client code. That can actually cause issues due to missing proc macro
hygiene.
Specifically, we have seen cases where a trait, let's call it Foo, is
implemented for any T, and it contains a very generic method such as
map. In such a scenario, the emitted initialization code could
result in an ambiguous method call, should it use, say, Iterator::map
somewhere. That is because we do not use fully qualified calling syntax
everywhere in generated code.
The ultimate fix should come from the Rust side in the form of proper
proc macro hygiene, but for now we prevent problems like this by putting
all initialization code into a generated module. The module boundary
ensures that only explicitly imported functions are available -- and we
do not import anything.

Closes #24