Build with minimum dependency versions
So far the testing we did for ensuring support for the stated minimum
supported Rust version has been accomplished by merely building the
crate with said version of the toolchain. In so doing, however, we
implicitly relied on the fact that all our dependencies (including
transitive ones) can in fact be built with this version of the
toolchain.
This is more and more unlikely, as crates increasingly use features from
newer versions of Rust or even entirely new editions.
With this change we decouple this check from those most recent
dependency versions. Rather, we build with the theoretically supported
minimum dependency versions.