Fix handling of test functions with arguments
We aim to support inner #[test] attributes. However, those attributes
are pretty much free to do whatever they like and transform the
user-written test in more or less arbitrary ways. Some of them may opt
to parametrize them with arguments and we cannot handle that currently,
as we assume a no-argument test function signature.
This change lifts this restriction, making it possible for us to support
constructs such as:

  #[test_log::test(test_case::test_case(-2, -4))]
  fn test_parametrized(x: i8, y: i8) {
    todo!()
  }

Note that it is likely that there are cases where we still fall short.
The test_case crate, for instance, supports the addition of test input
descriptions with syntax such as the following:

  #[test_case(-2, -4 ; "when both operands are negative")]

This construct is not a valid Rust function signature and it is probably
not possible to handle this and similar cases without special-casing for
each and every possible syntax. Similarly, generic arguments are not
handled at this point, but could conceivably be used by inner attributes.

Fixes: #27