Switch to using -Z direct-minimal-versions
So far we have relied on the unstable -Z minimal-versions switch to the
cargo update command in order to downgrade all dependencies to their
minimum supported versions to ensure that our own documented minimum
supported Rust version is kosher.
As it turns out this option can cause cargo update to take excessive
amounts of time and it's also unlikely to stabilize ever [0].
With this change we switch over to using -Z direct-minimal-versions
instead. This option only downgrades the direct dependencies of the
crate, which is a much less complex operation. As a cherry, doing so
allows us to get rid of the fake unused dependencies that we introduced
just for the sake of fixing upstream bugs.

[0] https://github.com/rust-lang/cargo/issues/12514