Put 'default_log_filter' attribute behind 'unstable' feature
The 'default_log_filter', while useful, is arguably a bit of a kludge
given that it is inherently racy with respect to other tests
initializing the corresponding log/tracing infrastructure beforehand.
What we would really want is to make it a global attribute, but that is
not currently possible on stable Rust.
Put this attribute behind a newly introduce feature, unstable, so that
we have more flexibility for changing/removing it in the future, without
being required to break semantic versioning guarantees.