Update publish workflow for test-log-macros
This change updates the publish workflow to accommodate for the recent
introduction of the test-log-macros crate to the workspace. Said crate
needs to be published before test-log itself, because test-log depends
on it. In terms of crate versioning we unconditionally release both
crates in lockstep for the sake of simplicity.