Stop using implicit dependency features
Our optional dependencies were implicitly defining features of the same
name, as that is how Cargo opted to do things. The behavior is weird and
confusing and probably best avoided. With version 1.60 of Cargo, we can
opt out by making actual features depend on said optional dependencies
via dep:<name> syntax [0].
With this change we do just that.

[0] https://doc.rust-lang.org/cargo/reference/features.html#optional-dependencies