Actually enable lints.rust usage
Turns out we had a typo in our workspace.lints.rust definition, causing
it to be ignored. Cargo spits out a warning, but there does not seem to
exist a way to convert that into a hard error, so that slipped through
CI.
Be that as it may, fix the typo and address the remaining "issues"
pointed out by the linter.