Emit tracing output to stderr
Behavior of the crate differs between the log and tracing backends:
while log defaults to printing messages to stderr, tracing emits them to
stdout. Until version 0.3.19 the tracing-subscriber crate did not allow
changing this default. Pull request #3187 [0] changed that and release
0.3.20 contains this logic.
Now that we can, adjust the tracing based backend to log to stderr for
consistency.

[0] https://github.com/tokio-rs/tracing/pull/3187