Use workspace-wide definitions for various meta-data
Cargo now allows us to define certain meta-data in a workspace-wide
manner and then more easily reuse that. We intend to update things such
as the edition for all crates in the workspace and the version of the
two crates being published from the repository should always be kept in
sync. This makes those properties (and others) perfect candidates for
this infrastructure.
With this change we switch over to using workspace-wide properties for a
bunch of meta-data.